+++ apps/windows-client/WorkshopOS.Client/ViewModels/ViewModels.cs
@@ -1,5 +1,5 @@
+using System.Collections.Concurrent;
 using System.Collections.ObjectModel;
-using System.Net;
 using System.Net.Http.Json;
 using System.Net.NetworkInformation;
 using System.Net.Sockets;